+++ src/ModelContextProtocol.Core/Server/McpServerFactory.cs
@@ -10,7 +10,7 @@ namespace ModelContextProtocol.Server;
 /// This is the recommended way to create <see cref="IMcpServer"/> instances.
 /// The factory handles proper initialization of server instances with the required dependencies.
 /// </remarks>
-[Obsolete($"Use {nameof(McpServer)}.{nameof(McpServer.Create)} instead.")]
+[Obsolete($"Use {nameof(McpServer)}.{nameof(McpServer.Create)} instead.")] // See: https://github.com/modelcontextprotocol/csharp-sdk/issues/774
 public static class McpServerFactory
 {
     /// <summary>
